Added cost as BigDecimal due to formatting issue

+++ src/main/kotlin/com/store/controllers/Products.kt
@@ -14,6 +14,7 @@ import org.springframework.web.bind.MethodArgumentNotValidException
 import com.fasterxml.jackson.annotation.JsonUnwrapped
 import java.time.LocalDateTime
 import java.time.format.DateTimeFormatter
+import java.math.BigDecimal
 import jakarta.servlet.http.HttpServletRequest
 import jakarta.validation.Valid
 import jakarta.validation.constraints.NotBlank
@@ -41,7 +42,7 @@ data class ProductDetails(
     @field:NotNull(message = "Cost is required")
     @field:DecimalMin(value = "0.01", message = "Cost must be at least 0.01")
     @field:DecimalMax(value = "999999.99", message = "Cost must be at most 999999.99")
-    val cost: Float 
+    val cost: BigDecimal 
 )
 
 data class ProductId(val id: Int)
